1. remove methods from Features module and add to each class 2. add "show" method to node_util
