+++ lib/cisco_node_utils.rb
@@ -15,7 +15,6 @@
 require 'cisco_node_utils/cisco_cmn_utils'
 require 'cisco_node_utils/command_reference'
 require 'cisco_node_utils/configparser_lib'
-require 'cisco_node_utils/features'
 require 'cisco_node_utils/interface'
 require 'cisco_node_utils/interface_ospf'
 require 'cisco_node_utils/node'
