+++ lib/cisco_node_utils/command_reference_common.yaml
@@ -211,7 +211,7 @@ interface:
     config_get: "show running interface all"
     config_get_token: ['/^interface %s$/i', '/^mtu (.*)$/']
     config_set: ["interface %s", "%s mtu %s"]
-    default_value: 1500
+    default_value: ""
 
   negotiate_auto_ethernet:
     config_get: "show running interface all"
